Good error information, fix program header checker.

+++ fileworking.h
@@ -1,3 +1,3 @@
 char* ReadFile (const char* filename);
 
-size_t GetFileSize (FILE* file);
+long GetFileSize (FILE* file);
